+++ HabitLoggerLibrary/Repository/HabitNotFoundException.cs
@@ -1,3 +1,3 @@
 namespace HabitLoggerLibrary.Repository;
 
-public sealed class HabitNotFoundException(int id) : Exception($"Habit with id: {id} was not found.");
+public sealed class HabitNotFoundException(long id) : Exception($"Habit with id: {id} was not found.");
